Move option stub code to a separate file
