+++ Funcky.Analyzers/Funcky.Analyzers.Test/OptionMatchAnalyzerTest.cs
@@ -6,60 +6,6 @@ namespace Funcky.Analyzers.Test;
 
 public sealed partial class OptionMatchAnalyzerTest
 {
-    private const string OptionCode =
-        """
-        namespace Funcky.Monads
-        {
-            public readonly struct Option<TItem>
-                where TItem : notnull
-            {
-                public static Option<TItem> None => default;
-
-                public static implicit operator Option<TItem>(TItem item) => default;
-
-                public TResult Match<TResult>(TResult none, System.Func<TItem, TResult> some) => default!;
-
-                public TResult Match<TResult>(System.Func<TResult> none, System.Func<TItem, TResult> some) => default!;
-
-                public TItem GetOrElse(TItem fallback) => default!;
-
-                public TItem GetOrElse(System.Func<TItem> fallback) => default!;
-
-                public Option<TItem> OrElse(Option<TItem> fallback) => default!;
-
-                public Option<TItem> OrElse(System.Func<Option<TItem>> fallback) => default!;
-
-                public Option<TResult> SelectMany<TResult>(System.Func<TItem, Option<TResult>> selector) where TResult : notnull => default;
-            }
-
-            public static class OptionExtensions
-            {
-                public static TItem? ToNullable<TItem>(this Option<TItem> option, RequireStruct<TItem>? ω = null) where TItem : struct => default;
-
-                public static TItem? ToNullable<TItem>(this Option<TItem> option, RequireClass<TItem>? ω = null) where TItem : class => default;
-            }
-
-            public static class Option
-            {
-                public static Option<TItem> Some<TItem>(TItem value) where TItem : notnull => default;
-
-                public static Option<TItem> Return<TItem>(TItem value) where TItem : notnull => default;
-            }
-
-            public sealed class RequireStruct<T> where T : struct { }
-
-            public sealed class RequireClass<T> where T : class { }
-        }
-
-        namespace Funcky
-        {
-            public static class Functional
-            {
-                public static T Identity<T>(T x) => x;
-            }
-        }
-        """;
-
     [Fact]
     public async Task IssuesWarningForReimplementationOfGetOrElse()
     {
@@ -114,7 +60,7 @@ public sealed partial class OptionMatchAnalyzerTest
             }
             """;
         await VerifyCS.VerifyCodeFixAsync(
-            inputCode + Environment.NewLine + OptionCode,
+            inputCode + Environment.NewLine + OptionStubCode,
             new[]
             {
                 VerifyCS.Diagnostic(PreferGetOrElse).WithSpan(10, 9, 10, 50),
@@ -124,7 +70,7 @@ public sealed partial class OptionMatchAnalyzerTest
                 VerifyCS.Diagnostic(PreferGetOrElse).WithSpan(14, 9, 14, 54),
                 VerifyCS.Diagnostic(PreferGetOrElse).WithSpan(20, 9, 20, 51),
             },
-            fixedCode + Environment.NewLine + OptionCode);
+            fixedCode + Environment.NewLine + OptionStubCode);
     }
 
     [Fact]
@@ -185,7 +131,7 @@ public sealed partial class OptionMatchAnalyzerTest
             }
             """;
         await VerifyCS.VerifyCodeFixAsync(
-            inputCode + Environment.NewLine + OptionCode,
+            inputCode + Environment.NewLine + OptionStubCode,
             new[]
             {
                 VerifyCS.Diagnostic(PreferOrElse).WithSpan(10, 9, 10, 71),
@@ -197,7 +143,7 @@ public sealed partial class OptionMatchAnalyzerTest
                 VerifyCS.Diagnostic(PreferOrElse).WithSpan(16, 9, 16, 69),
                 VerifyCS.Diagnostic(PreferOrElse).WithSpan(22, 9, 22, 58),
             },
-            fixedCode + Environment.NewLine + OptionCode);
+            fixedCode + Environment.NewLine + OptionStubCode);
     }
 
     [Fact]
@@ -248,14 +194,14 @@ public sealed partial class OptionMatchAnalyzerTest
             }
             """;
         await VerifyCS.VerifyCodeFixAsync(
-            inputCode + Environment.NewLine + OptionCode,
+            inputCode + Environment.NewLine + OptionStubCode,
             new[]
             {
                 VerifyCS.Diagnostic(PreferSelectMany).WithSpan(10, 9, 10, 84),
                 VerifyCS.Diagnostic(PreferSelectMany).WithSpan(11, 9, 11, 84),
                 VerifyCS.Diagnostic(PreferSelectMany).WithSpan(17, 9, 17, 63),
             },
-            fixedCode + Environment.NewLine + OptionCode);
+            fixedCode + Environment.NewLine + OptionStubCode);
     }
 
     [Fact]
@@ -288,6 +234,6 @@ public sealed partial class OptionMatchAnalyzerTest
                 }
             }
             """;
-        await VerifyCS.VerifyAnalyzerAsync(inputCode + Environment.NewLine + OptionCode);
+        await VerifyCS.VerifyAnalyzerAsync(inputCode + Environment.NewLine + OptionStubCode);
     }
 }
